requirements.txt

appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
boto3==1.17.69
botocore==1.20.69
brotlipy==0.7.0
certifi==2020.6.20
channels==3.0.3
colorama==0.4.4
conda==4.9.2
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.1.4
django-active-link==0.1.8
django-crispy-forms==1.10.0
django-crum==0.7.9
django-environ==0.4.5
django-extensions==3.1.2
django-filter==2.4.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-resized==0.3.11
django-simple-history==2.12.0
django-storages==1.11.1
filelock==3.0.12
flake8==3.8.4
graphviz==0.16
gunicorn==20.1.0
hyperlink==21.0.0
incremental==21.3.0
isort==5.6.4
jedi==0.17.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
menuinst==1.4.16
mysql-connector-python==8.0.23
parso==0.7.1
phonenumbers==8.12.18
Pillow==8.2.0
pluggy==0.13.1
protobuf==3.15.7
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.4
pywin32==227
rope==0.18.0
ruamel-yaml==0.15.87
s3transfer==0.4.2
service-identity==18.1.0
six 
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.52.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
ujson==4.0.1
validate-email==1.3
virtualenv==20.4.7
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.12.1
yapf==0.30.0
zope.interface==5.3.0
django-simple-history
django-crum
mysqlclient